Sneha, Imdad | #1756 | Fixed bugs found in dev box testing: operator should be retained when null, invalid grace count should default to null

+++ ananya-reports-web/src/test/java/org/motechproject/ananya/kilkari/reports/web/controller/SubscriptionControllerTest.java
@@ -88,6 +88,6 @@ public class SubscriptionControllerTest {
         SubscriptionStateChangeRequest subscriptionStateChangeRequest = subscriptionStateChangeRequestArgumentCaptor.getValue();
         assertEquals("sid", subscriptionStateChangeRequest.getSubscriptionId());
         assertEquals("ACTIVE", subscriptionStateChangeRequest.getSubscriptionStatus());
-        assertEquals(7, subscriptionStateChangeRequest.getGraceCount());
+        assertEquals((Integer) 7, subscriptionStateChangeRequest.getGraceCount());
     }
 }
